Code style and documentation update
1) Changed code style. Now private fields at the end of a class. Public fields after constuctors but before methods.
2) Documentation has been added for most classes, except for UI element classes.
3) Logger class renamed as Logging and changed its implementation.

+++ Sgl/Base/Threading/DelayAwaitable.cpp
@@ -3,7 +3,7 @@
 
 namespace Sgl
 {
-	DelayAwaitable::DelayAwaitable(size_t millisecondsDelay):
+	DelayAwaitable::DelayAwaitable(int64_t millisecondsDelay):
 		_duration(TimeSpan::FromMilliseconds(millisecondsDelay))
 	{}
 
@@ -11,7 +11,7 @@ namespace Sgl
 		_duration(delay)
 	{}
 
-	DelayAwaitable::DelayAwaitable(size_t millisecondsDelay, std::stop_token stopToken):
+	DelayAwaitable::DelayAwaitable(int64_t millisecondsDelay, std::stop_token stopToken):
 		_duration(TimeSpan::FromMilliseconds(millisecondsDelay)),
 		_stopToken(stopToken)
 	{}
